backend/requirements.txt
django==3.2
django-cors-headers==3.3.0
djangorestframework
psycopg2